move s3 api helpers into own TS file
