+++ scripts/binary/move-binaries.ts
@@ -1,10 +1,10 @@
+import { s3helpers } from './s3-api'
 const debug = require("debug")("cypress:binary")
 import la from 'lazy-ass'
 import is from 'check-more-types'
 // using "arg" module for parsing CLI arguments
 // because it plays really nicely with TypeScript
 import arg from 'arg'
-import S3 from 'aws-sdk/clients/s3'
 import {prop, sortBy, last} from 'ramda'
 import pluralize from 'pluralize'
 
@@ -96,88 +96,6 @@ export const prompts = {
     await confirm({
       question: 'Would you like to proceed? This will overwrite existing files',
       default: false,
-    })
-  }
-}
-
-/**
- * Utility object with methods that deal with S3.
- * Useful for testing our code that calls S3 methods.
- */
-export const s3helpers = {
-  makeS3 (aws) {
-    la(is.unemptyString(aws.key), 'missing aws key')
-    la(is.unemptyString(aws.secret), 'missing aws secret')
-
-    return new S3({
-      accessKeyId: aws.key,
-      secretAccessKey: aws.secret
-    })
-  },
-
-  verifyZipFileExists (zipFile: string, bucket: string, s3: S3): Promise<null> {
-    debug('checking S3 file %s', zipFile)
-    debug('bucket %s', bucket)
-
-    return new Promise((resolve, reject) => {
-      s3.headObject({
-        Bucket: bucket,
-        Key: zipFile
-      }, (err, data) => {
-        if (err) {
-          debug('error getting object %s', zipFile)
-          debug(err)
-
-          return reject(err)
-        }
-        debug('s3 data for %s', zipFile)
-        debug(data)
-        resolve()
-      })
-    })
-  },
-
-  /**
-   * Returns list of prefixes in a given folder
-   */
-  listS3Objects (uploadDir: string, bucket: string, s3: S3): Promise<string[]> {
-    la(is.unemptyString(uploadDir), 'invalid upload dir', uploadDir)
-
-    return new Promise((resolve, reject) => {
-      const prefix = uploadDir + '/'
-      s3.listObjectsV2({
-        Bucket: bucket,
-        Prefix: prefix,
-        Delimiter: '/'
-      }, (err, result) => {
-        if (err) {
-          return reject(err)
-        }
-
-        debug('AWS result in %s %s', bucket, prefix)
-        debug('%o', result)
-
-        resolve(result.CommonPrefixes.map(prop('Prefix')))
-      })
-    })
-  },
-
-  async copyS3 (sourceKey: string, destinationKey: string, bucket: string, s3: S3) {
-    return new Promise((resole, reject) => {
-      debug('copying %s in bucket %s to %s', sourceKey, bucket, destinationKey)
-
-      s3.copyObject({
-        Bucket: bucket,
-        CopySource: bucket + '/' + sourceKey,
-        Key: destinationKey
-      }, (err, data) => {
-        if (err) {
-          return reject(err)
-        }
-
-        debug('result of copying')
-        debug('%o', data)
-      })
     })
   }
 }
